添加消息队列来接收支付信息，同时更改异步提交订单和支付信息

+++ src/main/java/com/belong/model/Pay_Order.java
@@ -8,7 +8,7 @@ import java.io.Serializable;
 @Setter
 @Getter
 public class Pay_Order implements Serializable {
-    private Integer pay_id;
+    private Long pay_id;
     private Long order_id;
     private String pay_time;
     private String update_time;
